Add new endpoint worker status
