+++ plugins/worker/server/utils.py
@@ -17,11 +17,43 @@
 #  limitations under the License.
 ###############################################################################
 
+import celery
+
 from .constants import PluginSettings
 from girder.api.rest import getApiUrl
 from girder.models.file import File
 from girder.models.setting import Setting
+from girder.utility import setting_utilities
 from girder.plugins.jobs.models.job import Job
+
+_celeryapp = None
+
+
+def getCeleryApp():
+    """
+    Lazy loader for the celery app. Reloads anytime the settings are updated.
+    """
+    global _celeryapp
+
+    if _celeryapp is None:
+        backend = Setting().get(PluginSettings.BACKEND) or 'amqp://guest@localhost/'
+        broker = Setting().get(PluginSettings.BROKER) or 'amqp://guest@localhost/'
+        _celeryapp = celery.Celery('girder_worker', backend=backend, broker=broker)
+    return _celeryapp
+
+
+@setting_utilities.validator({
+    PluginSettings.BROKER,
+    PluginSettings.BACKEND
+})
+def validateSettings(doc):
+    """
+    Handle plugin-specific system settings. Right now we don't do any
+    validation for the broker or backend URL settings, but we do reinitialize
+    the celery app object with the new values.
+    """
+    global _celeryapp
+    _celeryapp = None
 
 
 def getWorkerApiUrl():
